+++ src/unix/solarish/mod.rs
@@ -513,9 +513,9 @@ s_no_extra_traits! {
 
     pub struct fd_set {
         #[cfg(target_pointer_width = "64")]
-        fds_bits: [i64; FD_SETSIZE / 64],
+        fds_bits: [i64; FD_SETSIZE as usize / 64],
         #[cfg(target_pointer_width = "32")]
-        fds_bits: [i32; FD_SETSIZE / 32],
+        fds_bits: [i32; FD_SETSIZE as usize / 32],
     }
 
     pub struct sockaddr_storage {
